delete by id

+++ src/main/java/com/songify/song/infrastructure/controller/SongMapper.java
@@ -27,7 +27,7 @@ public class SongMapper {
         return new CreateSongResponseDto(song);
     }
 
-    public static DeleteSongResponseDto mapFromSongToDeleteSongResponseDto(Integer id) {
+    public static DeleteSongResponseDto mapFromSongToDeleteSongResponseDto(Long id) {
         return new DeleteSongResponseDto("You deleted song with id: " + id, HttpStatus.OK);
     }
